add startnet.cmd

+++ Make_Minipe(ALL_IN_ONE).bat
@@ -32,24 +32,24 @@ for /f "delims=" %%i in (bin\Win10x86_64.txt) do (
         )
 )
 
-::修改注册表
-echo. & echo 准备释放注册表...... & echo.
-if not exist %~dp0build md %~dp0build
-bin\7z.exe e -o%~dp0build -aoa %wimfile% Windows/System32/config/system>NUL
-if "%Processor_Architecture%%Processor_Architew6432%" equ "x86" (
-set "NSudo=%~dp0x86\NSudo32.exe"
-) else (
-set "NSudo=%~dp0x64\NSudo64.exe"
-) 
-start "" /w /min %NSudo% -U:S -P:E -M:S "reg load hklm\minipe %~dp0build\system">NUL
-start "" /w /min %NSudo% -U:S -P:E -M:S "reg add "HKLM\minipe\ControlSet001\Services\mpssvc" /f /v "Start" /t REG_DWORD /d 3">NUL
-echo. & echo 挂载修改完毕，上载注册表...... & echo.
-start "" /w /min %NSudo% -U:S -P:E -M:S "reg unload hklm\minipe">nul
-start "" /w /min %NSudo% -U:S -P:E -M:S "reg unload hklm\minipe">nul
-echo. & echo 覆盖%wimfile%中的注册表...... & echo.
-bin\wimlib update %wimfile% --command="add '%~dp0build\system' '\Windows\System32\config\system'"
-::修改注册表完毕
-echo. & echo 正在增删削减%wimfile%包里的文件制作PE过程中，请您稍微等待 ...... & echo.
+:::||::挂载修改注册表方法备份
+:::||echo. & echo 准备释放注册表...... & echo.
+:::||if not exist %~dp0build md %~dp0build
+:::||bin\7z.exe e -o%~dp0build -aoa %wimfile% Windows/System32/config/system>NUL
+:::||if "%Processor_Architecture%%Processor_Architew6432%" equ "x86" (
+:::||set "NSudo=%~dp0x86\NSudo32.exe"
+:::||) else (
+:::||set "NSudo=%~dp0x64\NSudo64.exe"
+:::||) 
+:::||start "" /w /min %NSudo% -U:S -P:E -M:S "reg load hklm\minipe %~dp0build\system">NUL
+:::||start "" /w /min %NSudo% -U:S -P:E -M:S "reg add "HKLM\minipe\ControlSet001\Services\mpssvc" /f /v "Start" /t REG_DWORD /d 3">NUL
+:::||echo. & echo 挂载修改完毕，上载注册表...... & echo.
+:::||start "" /w /min %NSudo% -U:S -P:E -M:S "reg unload hklm\minipe">nul
+:::||start "" /w /min %NSudo% -U:S -P:E -M:S "reg unload hklm\minipe">nul
+:::||echo. & echo 覆盖%wimfile%中的注册表...... & echo.
+:::||bin\wimlib update %wimfile% --command="add '%~dp0build\system' '\Windows\System32\config\system'"
+:::||::修改注册表完毕
+  echo. & echo 正在增删削减%wimfile%包里的文件制作PE过程中，请您稍微等待 ...... & echo.
 bin\wimlib dir %wimfile% 1 --path=Windows\SysWOW64 | find ".exe" >NUL && (set FD=x64) || (set FD=x86)
 bin\wimlib update %wimfile%<excel.txt>NUL
 bin\wimlib update %wimfile%<%FD%\add2wim.txt>NUL
@@ -67,7 +67,7 @@ echo. & echo 结束时间：%endT%   耗时：%costT:~0,-2%.%costT:~-2% 秒
 del /f /q *.txt
 set output=%date:~0,4%%date:~5,2%%date:~8,2%%time:~0,2%%time:~3,2%%wimfile%
 ren %wimfile% "%output%"
-rd /s /q %~dp0build
+if exist %~dp0build rd /s /q %~dp0build
 echo. & echo 感谢您的等待，现在PE已经制作完成，%output%就是你的网络骨头版pe成品！ & echo.
 pause
 EXIT
